// frontend/src/app/types/stores.ts
export type Store = {
    storeId: string;
    storeCd: string;
    storeName: string;
    areaId: string;
    areaCd: string;
    areaName: string;
    storeTypeId: string;
    storeTypeCd: string;
    storeTypeName: string;
    isActive: boolean;
    updatedAt: string;
    updatedBy: string;
    createdAt: string;
    createdBy: string;
  };